Restore more zizmor-audit APIs
