+++ crates/zizmor-audit/src/lib.rs
@@ -1,8 +1,10 @@
 #![warn(clippy::all, clippy::dbg_macro)]
 
-use zizmor_core::{github, input, models, utils};
+use zizmor_core::{github, input, models};
 
 pub mod audit;
 pub mod finding;
 pub mod registry;
 pub mod state;
+mod uses;
+mod utils;
