+++ crates/zizmor-core/Cargo.toml
@@ -31,9 +31,6 @@ subfeature.workspace = true
 terminal-link.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
-tree-sitter.workspace = true
-tree-sitter-bash.workspace = true
-tree-sitter-powershell.workspace = true
 url.workspace = true
 yaml_serde.workspace = true
 yamlpath.workspace = true
